+++ Sora/Sora.swift
@@ -153,6 +153,14 @@ public final class Sora: @unchecked Sendable {
         _ error: Error?
       ) -> Void
   ) -> ConnectionTask {
+    if let error = configureBeforeConnect(configuration: configuration) {
+      let task = ConnectionTask()
+      task.complete()
+      handler(nil, error)
+      handlers.onConnect?(nil, error)
+      return task
+    }
+
     let mediaChan = MediaChannel(manager: self, configuration: configuration)
     mediaChan.internalHandlers.onDisconnectLegacy = { [weak self, weak mediaChan] error in
       guard let weakSelf = self else {
@@ -188,6 +196,20 @@ public final class Sora: @unchecked Sendable {
       handler(mediaChan, nil)
       weakSelf.handlers.onConnect?(mediaChan, nil)
     }
+  }
+
+  // Configuration の組み合わせチェックと、接続前の事前適用を行います
+  func configureBeforeConnect(configuration: Configuration) -> Error? {
+    // customAudioDevice 利用時は AudioInputMuteControllable プロトコル準拠となっている必要があります
+    if let customAudioDevice = configuration.customAudioDevice,
+      !(customAudioDevice is AudioInputMuteControllable)
+    {
+      return SoraError.mediaChannelError(
+        reason: "customAudioDevice must conform to AudioInputMuteControllable when sending audio"
+      )
+    }
+
+    return nil
   }
 
   // MARK: - 音声ユニットの操作